Updating types.
* Changing `np.float32` to `float`.

+++ examples/reach/stoch_reach.py
@@ -70,11 +70,11 @@ S = transition_sampler(env, state_sampler, action_sampler).sample(size=sample_si
 
 # %%
 target_tube = [
-    gym.spaces.Box(low=-0.5, high=0.5, shape=(2,), dtype=np.float32)
+    gym.spaces.Box(low=-0.5, high=0.5, shape=(2,), dtype=float)
 ] * time_horizon
 
 constraint_tube = [
-    gym.spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
+    gym.spaces.Box(low=-1, high=1, shape=(2,), dtype=float)
 ] * time_horizon
 
 # Generate test points.
